refactor: split auth domain from user
