+++ app/main.py
@@ -39,7 +39,7 @@ app.add_middleware(
 
 # 라우터 등록
 app.include_router(user.router, prefix="/api/v1/users", tags=["Users"])
-app.include_router(auth.router, prefix="/api/v1/oauth2", tags=["OAuth2"])
+app.include_router(auth.router, prefix="/api/v1/auth", tags=["Auth"])
 app.include_router(problem.router, prefix="/api/v1/problems", tags=["Problems"])
 app.include_router(category.router, prefix="/api/v1/categories", tags=["Categories"])
 app.include_router(embedding.router, prefix="/api/v1/embedding", tags=["Embedding"])
